Add support to add terraform and Ansible code (#14)
* Adding Ansible incoming repo

* Adding Ansible extra-vars from deployment repo

* Change calling repo path

* Fixing some inception paths

* Adding protocol option to elb

* Fixing EC2 var for port exposure

* Fixing some EFS checks - Adding ansible skip

* Removing unnecesary Ansible declarations

* Spliting incoming repo's script into another bash

* Added root volume size option

* Added instance protect and ami update ignore

* Adding env_bitops option

* Ami lifecycle to generator - Fix docker extra-vars

* Forcing ansible skip when tf_stack_destroy is set

* Moved up Terraform run-type validation check

* Adding root_vol_preserve

* Adding DB protection and final snapshot

* Adding support for EFS deletion protection through action

* Adding error message for invalid instance type in terraform

* Adding count to az_select to avoid error

* Generated code storage

* Adding Code-only option

+++ operations/_scripts/deploy/deploy.sh
@@ -2,8 +2,43 @@
 
 #set -x
 
+# Removes anything from the variable and leave only alpha characters, and lowers them. This is to validate if boolean.
+function alpha_only() {
+    echo "$1" | tr -cd '[:alpha:]' | tr '[:upper:]' '[:lower:]'
+}
+
 echo "::group::In Deploy"
 GITHUB_REPO_NAME=$(echo $GITHUB_REPOSITORY | sed 's/^.*\///')
+
+# Validating if Terraform is set to destroy, and avoid Ansible
+TERRAFORM_COMMAND=""
+if [ "$(alpha_only $TF_STACK_DESTROY)" == "true" ]; then
+  TERRAFORM_COMMAND="destroy"
+  ANSIBLE_SKIP="true"
+  if [ "$(alpha_only $AWS_EC2_INSTANCE_PROTECT)" == "true" ] && [ "$(alpha_only $AWS_EC2_INSTANCE_CREATE)" == "true" ]; then
+    echo "::error::You need to set aws_ec2_instance_protect to false before before destroying infrastructure."
+    exit 1
+  fi
+  if [ "$(alpha_only $AWS_POSTGRES_DATABASE_PROTECTION)" == "true" ]; then
+    echo "::error::Database protection enabled. Disable it before destroying."
+    exit 1
+  fi
+  if [ "$(alpha_only $AWS_EFS_VOLUME_PRESERVE)" == "true" ]; then
+    echo "::notice::There is no real EFS protection to enable. Just a flag we created to avoid unintentional deletion."
+    echo "::notice::Please backup your volume before deletion."
+    echo "::error::EFS volume protection enabled. Disable it before destroying."
+    exit 1
+  fi
+else 
+  if [ "$(alpha_only $AWS_EFS_VOLUME_PRESERVE)" == "true" ]; then
+    echo "::notice::There is no real EFS protection to enable from AWS."
+    echo "::notice::This is just a flag we created to avoid unintentional deletion on destruction."
+  fi
+fi
+
+if [ "$(alpha_only $ANSIBLE_SKIP)" == "true" ]; then
+  ANSIBLE_SKIP="true"
+fi
 
 # Generate buckets identifiers and check them agains AWS Rules 
 export TF_STATE_BUCKET="$(/bin/bash $GITHUB_ACTION_PATH/operations/_scripts/generate/generate_buckets_identifiers.sh tf | xargs)"
@@ -25,36 +60,29 @@ export LB_LOGS_BUCKET="$(/bin/bash $GITHUB_ACTION_PATH/operations/_scripts/gener
 # Generate bitops config
 /bin/bash $GITHUB_ACTION_PATH/operations/_scripts/generate/generate_bitops_config.sh
 
+# Generate bitops incoming repos config
+if [ -n "$GH_ACTION_REPO" ]; then
+  if [ -n "$GH_ACTION_INPUT_TERRAFORM" ] || [ -n "$GH_ACTION_INPUT_ANSIBLE" ]; then
+    /bin/bash $GITHUB_ACTION_PATH/operations/_scripts/generate/generate_bitops_incoming.sh
+  fi
+  # Generating incoming extra_vars_file if it exists
+  if [ -n "$BITOPS_EXTRA_ENV_VARS_FILE" ]; then
+    if [ -s $GH_ACTION_REPO/$BITOPS_EXTRA_ENV_VARS_FILE ]; then
+      BITOPS_EXTRA_ENV_VARS_FILE="--env-file $GH_ACTION_REPO/$BITOPS_EXTRA_ENV_VARS_FILE"
+      cat $GH_ACTION_REPO/$BITOPS_EXTRA_ENV_VARS_FILE
+    else
+      echo "File $BITOPS_EXTRA_ENV_VARS_FILE missing or empty"
+    fi
+  fi
+fi
+
 # Generating GitHub Variables and Secrets files
 mkdir -p "${GITHUB_ACTION_PATH}/operations/deployment/env-files"
 echo "$ENV_GHV" > "${GITHUB_ACTION_PATH}/operations/deployment/env-files/ghv.env"
 echo "$ENV_GHS" > "${GITHUB_ACTION_PATH}/operations/deployment/env-files/ghs.env"
-if [ -s "$GITHUB_WORKSPACE/$ENV_REPO" ]; then
+if [ -s "$GITHUB_WORKSPACE/$ENV_REPO" ] && [ -n "$ENV_REPO" ]; then
   cp "$GITHUB_WORKSPACE/$ENV_REPO" "${GITHUB_ACTION_PATH}/operations/deployment/env-files/repo.env"
 fi
-
-
-# DEBUGGING --- TBD
-
-# List terraform folder
-echo "ls -al $GITHUB_ACTION_PATH/operations/deployment/terraform/"
-ls -al $GITHUB_ACTION_PATH/operations/deployment/terraform/
-# Prints out bitops.config.yaml
-echo "cat $GITHUB_ACTION_PATH/operations/deployment/terraform/bitops.config.yaml"
-cat $GITHUB_ACTION_PATH/operations/deployment/terraform/bitops.config.yaml
-
-echo "cat GITHUB_ACTION_PATH/operations/deployment/terraform/provider.tf"
-cat $GITHUB_ACTION_PATH/operations/deployment/terraform/provider.tf
-echo "ls GITHUB_ACTION_PATH/operations/deployment/docker/app/${GITHUB_REPO_NAME}"
-ls "$GITHUB_ACTION_PATH/operations/deployment/docker/app/${GITHUB_REPO_NAME}"
-
-
-TERRAFORM_COMMAND=""
-if [ "$TF_STACK_DESTROY" == "true" ]; then
-  TERRAFORM_COMMAND="destroy"
-  ANSIBLE_SKIP_DEPLOY="true"
-fi
-echo "::endgroup::"
 
 if [[ $SKIP_BITOPS_RUN == "true" ]]; then
   exit 1
@@ -71,14 +99,23 @@ docker run --rm --name bitops \
 -e SKIP_DEPLOY_TERRAFORM="${SKIP_DEPLOY_TERRAFORM}" \
 -e SKIP_DEPLOY_HELM="${SKIP_DEPLOY_HELM}" \
 -e BITOPS_TERRAFORM_COMMAND="${TERRAFORM_COMMAND}" \
--e BITOPS_ANSIBLE_SKIP_DEPLOY="${ANSIBLE_SKIP_DEPLOY}" \
+-e BITOPS_ANSIBLE_SKIP_DEPLOY="${ANSIBLE_SKIP}" \
 -e TF_STATE_BUCKET="${TF_STATE_BUCKET}" \
 -e TF_STATE_BUCKET_DESTROY="${TF_STATE_BUCKET_DESTROY}" \
 -e DEFAULT_FOLDER_NAME="_default" \
 -e BITOPS_FAST_FAIL="${BITOPS_FAST_FAIL}" \
+${BITOPS_EXTRA_ENV_VARS_FILE} \
+${BITOPS_EXTRA_ENV_VARS} \
 -v $(echo $GITHUB_ACTION_PATH)/operations:/opt/bitops_deployment \
-bitovi/bitops:dev
+bitovi/bitops:2.5.0
 BITOPS_RESULT=$?
 echo "::endgroup::"
+
+if [[ "$(alpha_only $BITOPS_CODE_ONLY)" == "true" ]]; then
+  echo "::group::Generating BitOps code archive"  
+  echo "Generating bitops.config.yaml based on your definitions:"
+  /bin/bash $GITHUB_ACTION_PATH/operations/_scripts/generate/generate_bitops_config_code_only.sh
+  echo "::endgroup::"
+fi
 
 exit $BITOPS_RESULT
